These tables may already exist, don't overwrite them in that case.

+++ db/migrate/20100913234707_create_refinerycms_images_schema.rb
@@ -10,14 +10,14 @@ class CreateRefinerycmsImagesSchema < ActiveRecord::Migration
       t.datetime "updated_at"
       t.string   "image_uid"
       t.string   "image_ext"
-    end
+    end unless ::Image.table_exists?
   end
 
   def self.down
     [::Image].reject{|m|
       !(defined?(m) and m.respond_to?(:table_name))
     }.each do |model|
-      drop_table model.table_name
+      drop_table model.table_name if model.table_exists?
     end
   end
 end
