+++ db/migrate/20100913234709_create_refinerycms_resources_schema.rb
@@ -8,14 +8,14 @@ class CreateRefinerycmsResourcesSchema < ActiveRecord::Migration
       t.datetime "updated_at"
       t.string   "file_uid"
       t.string   "file_ext"
-    end
+    end unless ::Resource.table_exists?
   end
 
   def self.down
     [::Resource].reject{|m|
       !(defined?(m) and m.respond_to?(:table_name))
     }.each do |model|
-      drop_table model.table_name
+      drop_table model.table_name if model.table_exists?
     end
   end
 end
